// src/api/generated/model/firmographicConditionField.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */

export type FirmographicConditionField = typeof FirmographicConditionField[keyof typeof FirmographicConditionField];


export const FirmographicConditionField = {
  employees: 'employees',
  revenue_eur: 'revenue_eur',
  country_code: 'country_code',
  industry_ids: 'industry_ids',
  domain: 'domain',
  tags: 'tags',
} as const;